+++ docs/source/Multi-Modal/llava-video最佳实践.md
@@ -93,7 +93,7 @@ print()
 
 """
 query: <video>描述这段视频
-response: 这段视频展示了一个小孩在床上享受一本书的愉悦。Ъ_Ъ。她穿着一件简单的纱衣，头戴着眼镜，手轻轻地摸索着书页。她的眼神在书中漫游，她的脸上带着一种快乐的表情，仿佛在感受到故事的情感。她的手臂在书上滑动，她的脚在床上踩着，仿佛在等待故事的下一个段落。这个瞬间充满了快乐和温馨，展示了小孩在阅读中的满足和愉悦。
+response: 这段视频展示了一个小孩在床上享受一本书的愉悦。她穿着一件简单的纱衣，头戴着眼镜，手轻轻地摸索着书页。她的表情充满了兴奋和惊喜，她的眼睛时不时地眨眼地看着书页，仿佛在探索一个新的世界。她的姿势和动作都充满了轻松和自然，让人感觉到她在享受这个简单而美好的时刻。
 query: <image>图中有几只羊
 response: 在这张图像中，有四只羊。
 """
